vejr-db/requirements.txt
requests
psycopg2-binary
pandas
matplotlib